+++ src/pland/hooks/Hook.cc
@@ -21,7 +21,7 @@
 #include "mc/world/level/BlockSource.h"
 #include "mc/world/level/Level.h"
 #include "mc/world/level/block/FireBlock.h"
-
+#include "mc/world/level/block/actor/ChestBlockActor.h"
 namespace land {
 
 
@@ -64,20 +64,23 @@ LL_TYPE_INSTANCE_HOOK(
     auto* db   = PLand::getInstance().getLandRegistry();
     auto  land = db->getLandAt(pos, dimId);
 
+    auto* player = hookActor.getPlayerOwner();
+    if (!player) {
+        origin(inEntity, inSpeed);
+        return;
+    }
+
     // 如果在领地内
     if (land) {
-        // 检查 inEntity 是否为玩家
-        if (inEntity.isPlayer()) {
-            auto& player = static_cast<Player&>(inEntity);
-            if (!PreCheckLandExistsAndPermission(land, player.getUuid())) {
-                // 领地不存在或玩家没有权限，则拦截
-                return;
-            }
-            // 检查钓鱼竿权限
-            if (!land->getPermTable().allowFishingRodAndHook) {
-                // 如果不允许使用钓鱼竿，则拦截
-                return;
-            }
+        // 检查玩家是否有权限
+        if (!PreCheckLandExistsAndPermission(land, player->getUuid())) {
+            // 领地不存在或玩家没有权限，则拦截
+            return;
+        }
+        // 检查钓鱼竿权限
+        if (!land->getPermTable().allowFishingRodAndHook) {
+            // 如果不允许使用钓鱼竿，则拦截
+            return;
         }
     }
     origin(inEntity, inSpeed);
@@ -115,7 +118,8 @@ LL_TYPE_INSTANCE_HOOK(
     ::BlockPos const& pos,
     int               chance,
     ::Randomize&      randomize,
-    int               age
+    int               age,
+    ::BlockPos const& firePos
 ) {
     // 获取领地注册表实例
     auto* db   = PLand::getInstance().getLandRegistry();
@@ -125,6 +129,33 @@ LL_TYPE_INSTANCE_HOOK(
     if (land && !land->getPermTable().allowFireSpread) {
         return;
     }
+    origin(region, pos, chance, randomize, age, firePos);
+}
+
+
+
+// Fix [#158](https://github.com/engsr6982/PLand/issues/158)
+LL_TYPE_INSTANCE_HOOK(
+    ChestBlockActorOpenHook,
+    ll::memory::HookPriority::Normal,
+    ChestBlockActor,
+    &ChestBlockActor::$startOpen,
+    void,
+    ::Actor& actor
+) {
+    if (actor.isPlayer()) {
+        origin(actor);
+        return;
+    }
+    // 获取领地注册表实例
+    auto* db   = PLand::getInstance().getLandRegistry();
+    auto  land = db->getLandAt(this->mPosition, actor.getDimensionId());
+
+
+    if (land && !land->getPermTable().allowOpenChest) {
+        return;
+    }
+    origin(actor);
 }
 
 // impl EventListener
@@ -133,6 +164,7 @@ void EventListener::registerHooks() {
     RegisterHookIf<FishingHookHitHook>(Config::cfg.hooks.registerFishingHookHitHook);
     RegisterHookIf<LayEggGoalHook>(Config::cfg.hooks.registerLayEggGoalHook);
     RegisterHookIf<FireBlockBurnHook>(Config::cfg.hooks.registerFireBlockBurnHook);
+    RegisterHookIf<ChestBlockActorOpenHook>(Config::cfg.hooks.registerChestBlockActorOpenHook);
 }