+++ src/pland/infra/Config.h
@@ -18,7 +18,7 @@ struct ForbiddenRange {
 };
 
 struct Config {
-    int              version{30};
+    int              version{31};
     ll::io::LogLevel logLevel{ll::io::LogLevel::Info};
 
     EconomyConfig economy;
@@ -121,6 +121,7 @@ struct Config {
         bool registerFishingHookHitHook{true}; // 注册钓鱼竿Hook
         bool registerLayEggGoalHook{true};     // 注册产卵AI目标Hook
         bool registerFireBlockBurnHook{true};  // 注册火焰蔓延Hook
+        bool registerChestBlockActorOpenHook{true}; // 注册箱子打开Hook
     } hooks;
 
     struct {